Add a mode to prevent polling the master state.
If we poll the master's `/state` endpoint on large clusters, we can
degrade master performance. Since the `/state` endpoint is used to
generate per-agent resource metrics, and we can already get those
by running the exporter on each agent, add a new flag that can
disable the collection of these metrics.